TopSellingProduct: TotalQuantity property type is adjusted from int to long to avoid possible overflow.

+++ src/ChannelEngineTopSellingProducts.Application/Products/TopSellingProductDto.cs
@@ -5,5 +5,5 @@ public sealed class TopSellingProductDto
 	public string Id { get; init; } = string.Empty;
 	public string Name { get; init; } = string.Empty;
 	public string Gtin { get; init; } = string.Empty;
-	public int TotalQuantity { get; init; }
+	public long TotalQuantity { get; init; }
 }
